# Conflicts:
#	TeensyEngine.ino

+++ src/Geometry.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Vec3.h"
+#include <vector>
 
 struct Vertex {
 	Vertex() {}
@@ -82,12 +83,12 @@ template<class T>
 class IndexedTriangleList {
 public:
 	IndexedTriangleList() = default;
-	IndexedTriangleList(std::vector<T> verts_in, std::vector<size_t> indices_in) :
+	IndexedTriangleList(std::vector<T> verts_in, std::vector<uint16_t> indices_in) :
 		vertices(std::move(verts_in)),
 		indices(std::move(indices_in)) {}
 
 	std::vector<T> vertices;
-	std::vector<size_t> indices;
+	std::vector<uint16_t> indices;
 };
 
 template<class V>
